Conditionally compile classes used in DEBUG only
